pyproject.toml
[project]
name = "product-impact-analizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Victor Camargo", email = "victor.camargo@alliesystems.com" }
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "openai>=1.0.0",
    "langfuse>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
